[skip ai review] refactor: split PanelMainView into +Content and +Banners extensions
