+++ Sources/RunBot/Views/Main/PanelMainView+Content.swift
@@ -1,0 +1,69 @@
+// PanelMainView+Content.swift
+// RunBot
+//
+// Workflow content section and active-local-runner derivation for PanelMainView.
+// Extracted from PanelMainView.swift for readability — no behaviour changes.
+// All sizing contract invariants live in PanelMainView.swift.
+import GitHubClient
+import RunBotCore
+import SwiftUI
+
+extension PanelMainView {
+
+    // MARK: - Active local runners
+
+    /// Local runners currently executing a job inside an in-progress workflow group.
+    ///
+    /// Reads GitHub-side state (`actions`, `jobs`, `runners`) and local runner state
+    /// (`localRunners`) from `runnerState` — the single observable source of truth
+    /// injected via the SwiftUI environment from `AppDelegate.wrapEnv`.
+    var activeLocalRunners: [RunnerModel] {
+        guard appState.runnerState.actions.contains(where: { $0.groupStatus == .inProgress }) else { return [] }
+        let activeNamesFromJobs = Set(
+            appState.runnerState.jobs.filter { $0.jobStatus == .inProgress }.compactMap { $0.runnerName }
+        )
+        let busyRunners = appState.runnerState.runners.filter { $0.busy }
+        let busyIds = Set(busyRunners.compactMap { $0.id })
+        let busyNames = Set(busyRunners.map { $0.name })
+        return appState.runnerState.localRunners.filter { local in
+            if activeNamesFromJobs.contains(local.runnerName) { return true }
+            if let aid = local.agentId, busyIds.contains(aid) { return true }
+            if busyNames.contains(local.runnerName) { return true }
+            return false
+        }
+    }
+
+    // MARK: - Content
+
+    /// Workflow rows and the load-more button, rendered inside the scroll container.
+    var actionsSectionContent: some View {
+        VStack(alignment: .leading, spacing: 0) {
+            SectionHeaderLabel(title: "Workflows")
+            if appState.runnerState.actions.isEmpty {
+                Text("No recent workflows")
+                    .font(.caption).foregroundColor(.secondary)
+                    .padding(.horizontal, 12).padding(.vertical, 8)
+            } else {
+                let visible = Array(appState.runnerState.actions.prefix(visibleCount))
+                ForEach(visible) { group in
+                    ActionRowView(group: group, tick: displayTick, onStepTap: onStepTap)
+                }
+                loadMoreButton
+            }
+        }
+        .padding(.vertical, 4)
+    }
+
+    /// "Load N more workflows" button; hidden when all workflows are already visible.
+    @ViewBuilder var loadMoreButton: some View {
+        let nextBatch = min(10, appState.runnerState.actions.count - visibleCount)
+        if nextBatch > 0 {
+            Button { visibleCount += nextBatch } label: {
+                Text("Load \(nextBatch) more workflows\u{2026}")
+                    .font(.caption).foregroundColor(.secondary)
+            }
+            .buttonStyle(.plain)
+            .padding(.horizontal, 12).padding(.vertical, 6)
+        }
+    }
+}
